+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/SearchMenuManager.java
@@ -9,9 +9,10 @@ import gov.healthit.chpl.domain.DescriptiveModel;
 import gov.healthit.chpl.domain.KeyValueModel;
 import gov.healthit.chpl.domain.KeyValueModelStatuses;
 import gov.healthit.chpl.domain.SurveillanceRequirementOptions;
+import gov.healthit.chpl.domain.notification.NotificationType;
 
 public interface SearchMenuManager {
-	
+	public Set<NotificationType> getNotificationTypes();
 	public Set<KeyValueModel> getClassificationNames();
 	public Set<KeyValueModel> getEditionNames(Boolean simple);
 	public Set<KeyValueModel> getCertificationStatuses();